requirements.txt
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
boto3==1.24.84
botocore==1.27.84
certifi==2022.9.24
charset-normalizer==2.1.1
debugpy==1.6.3
decorator==5.1.1
display==1.0.0
docopt==0.6.2
entrypoints==0.4
executing==1.1.0
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jmespath==1.0.1
jupyter-client==7.3.5
jupyter-core==4.11.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.11
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
stack-data==0.5.1
toml==0.10.2
tornado==6.2
traitlets==5.4.0
tweepy==4.10.1
twitter==1.19.6
TwitterSearch==1.0.2
upgrade-requirements==1.7.0
urllib3==1.26.12
wcwidth==0.2.5
yarg==0.1.9
